Fixing build on linux
